Early rld external-fixup work. Relocation addend is signed.

+++ llvm/lib/MC/RepoObjectWriter.cpp
@@ -277,7 +277,8 @@ void RepoObjectWriter::recordRelocation(MCAssembler &Asm,
   }
 
   Relocations[&FixupSection].emplace_back(FixupOffset, RenamedSymA, Type,
-                                          Addend, SymA, C, UsedSymbolName);
+                                          static_cast<std::int64_t>(Addend),
+                                          SymA, C, UsedSymbolName);
 }
 
 namespace {
